Moved throwing of the exception to IllegalOperationChecker.
