+++ rxandroidble/src/main/java/com/polidea/rxandroidble/BluetoothGattCharacteristicProperty.java
@@ -1,0 +1,3 @@
+package com.polidea.rxandroidble;
+
+public @interface BluetoothGattCharacteristicProperty { }
