+++ rxandroidble/src/main/java/com/polidea/rxandroidble/exceptions/BleIllegalOperationException.java
@@ -2,6 +2,8 @@ package com.polidea.rxandroidble.exceptions;
 
 import android.support.annotation.RestrictTo;
 
+import com.polidea.rxandroidble.BluetoothGattCharacteristicProperty;
+
 import java.util.UUID;
 
 /**
@@ -11,8 +13,8 @@ import java.util.UUID;
 public class BleIllegalOperationException extends RuntimeException {
 
     public final UUID characteristicUUID;
-    public final int supportedProperties;
-    public final int neededProperties;
+    public final @BluetoothGattCharacteristicProperty int supportedProperties;
+    public final @BluetoothGattCharacteristicProperty int neededProperties;
 
     @RestrictTo(RestrictTo.Scope.LIBRARY_GROUP)
     public BleIllegalOperationException(String message,
