+++ rxandroidble/src/main/java/com/polidea/rxandroidble/internal/connection/IllegalOperationHandler.java
@@ -3,6 +3,8 @@ package com.polidea.rxandroidble.internal.connection;
 
 import android.bluetooth.BluetoothGattCharacteristic;
 
+import com.polidea.rxandroidble.exceptions.BleIllegalOperationException;
+
 /**
  * Handler for {@link IllegalOperationChecker#checkAnyPropertyMatches(BluetoothGattCharacteristic, int)} response.
  */
@@ -14,5 +16,5 @@ public abstract class IllegalOperationHandler {
         this.messageCreator = messageCreator;
     }
 
-    public abstract void handleMismatchData(BluetoothGattCharacteristic characteristic, int neededProperties);
+    public abstract BleIllegalOperationException handleMismatchData(BluetoothGattCharacteristic characteristic, int neededProperties);
 }
