+++ rxandroidble/src/main/java/com/polidea/rxandroidble/internal/connection/ThrowingIllegalOperationHandler.java
@@ -8,7 +8,7 @@ import com.polidea.rxandroidble.internal.RxBleLog;
 import javax.inject.Inject;
 
 /**
- * Implementation of {@link IllegalOperationHandler}. This class logs an error and throws {@link BleIllegalOperationException} if there
+ * Implementation of {@link IllegalOperationHandler}. This class logs an error and returns {@link BleIllegalOperationException} if there
  * was no match between possessed and requested properties.
  */
 public class ThrowingIllegalOperationHandler extends IllegalOperationHandler {
@@ -19,15 +19,15 @@ public class ThrowingIllegalOperationHandler extends IllegalOperationHandler {
     }
 
     /**
-     * This method logs an error and throws a {@link BleIllegalOperationException}.
+     * This method logs an error and returns a {@link BleIllegalOperationException}.
      * @param characteristic the characteristic upon which the operation was requested
      * @param neededProperties bitmask of properties needed by the operation
      */
     @Override
-    public void handleMismatchData(BluetoothGattCharacteristic characteristic, int neededProperties) {
+    public BleIllegalOperationException handleMismatchData(BluetoothGattCharacteristic characteristic, int neededProperties) {
         String message = messageCreator.createMismatchMessage(characteristic, neededProperties);
         RxBleLog.e(message);
-        throw new BleIllegalOperationException(message,
+        return new BleIllegalOperationException(message,
                 characteristic.getUuid(),
                 characteristic.getProperties(),
                 neededProperties);
