+++ rxandroidble/src/test/groovy/com/polidea/rxandroidble/internal/connection/IllegalOperationCheckerTest.groovy
@@ -1,8 +1,10 @@
-package com.polidea.rxandroidble.internal.util
+package com.polidea.rxandroidble.internal.connection
 
 import android.bluetooth.BluetoothGattCharacteristic
+import com.polidea.rxandroidble.exceptions.BleIllegalOperationException
 import com.polidea.rxandroidble.internal.connection.IllegalOperationChecker
 import com.polidea.rxandroidble.internal.connection.IllegalOperationHandler
+import rx.observers.TestSubscriber
 import spock.lang.Specification
 
 class IllegalOperationCheckerTest extends Specification {
@@ -10,25 +12,29 @@ class IllegalOperationCheckerTest extends Specification {
     IllegalOperationChecker objectUnderTest
     BluetoothGattCharacteristic bluetoothGattCharacteristic
     IllegalOperationHandler mockHandler
+    BleIllegalOperationException mockException
+    TestSubscriber testSubscriber
     final int NO_PROPERTIES = 0
 
     void setup() {
         mockHandler = Mock IllegalOperationHandler
         objectUnderTest = new IllegalOperationChecker(mockHandler)
-
+        mockException = Mock BleIllegalOperationException
         bluetoothGattCharacteristic = Mock BluetoothGattCharacteristic
+        testSubscriber = new TestSubscriber()
     }
 
     def "should handle message if no property matches"() {
         given:
         bluetoothGattCharacteristic.getProperties() >> NO_PROPERTIES
+        mockHandler.handleMismatchData(_, _) >> mockException
 
         when:
         objectUnderTest.checkAnyPropertyMatches(bluetoothGattCharacteristic, BluetoothGattCharacteristic.PROPERTY_READ)
-                .subscribe()
+                .subscribe(testSubscriber)
 
         then:
-        1 * mockHandler.handleMismatchData(_, _)
+        testSubscriber.assertError(mockException)
     }
 
     def "should not handle message if the only needed property matches"() {
@@ -37,10 +43,10 @@ class IllegalOperationCheckerTest extends Specification {
 
         when:
         objectUnderTest.checkAnyPropertyMatches(bluetoothGattCharacteristic, BluetoothGattCharacteristic.PROPERTY_READ)
-                .subscribe()
+                .subscribe(testSubscriber)
 
         then:
-        0 * mockHandler.handleMismatchData(_, _)
+        testSubscriber.assertCompleted()
     }
 
     def "should not handle message if at least one of the needed properties match"() {
@@ -52,9 +58,9 @@ class IllegalOperationCheckerTest extends Specification {
                 BluetoothGattCharacteristic.PROPERTY_WRITE
                         | BluetoothGattCharacteristic.PROPERTY_SIGNED_WRITE
                         | BluetoothGattCharacteristic.PROPERTY_WRITE_NO_RESPONSE)
-                .subscribe()
+                .subscribe(testSubscriber)
 
         then:
-        0 * mockHandler.handleMismatchData(_, _)
+        testSubscriber.assertCompleted()
     }
 }
